Project Update
- Updated gitignore file so that it does not ignore the precompiled
libraries or binary files for windows, but did specify to ignore the
root build directory

- This is technically the first commit so please allow me to explain: 

I first created the connector and had it working just find using both
statements and prepared statements (simple select/join/limit/orderby),
and doing advanced query manipulation and string composition.

Then I started adding facilities for integrating with a 'model' system
to provide a seamless feeling work flow for the developer (NodaKwaeri).

I then switched over to windows to make sure everything worked there as
well (this was after developing for debian 7.5 64 bit).  There were
issues ofc so I decided to take the MySQL C++ Connector and create a
Node-Gyp build process for creating the shared library on both windows
and linux to minify as many manual steps as possible.

Currently I've gotten the build process mostly complete on windows,
there are some errors that I'm dealing with.  You'll need to have MySQL
Server 5.6.* (19 at the time of this checkin) 64-bit or 32-bit installed
depending on the node.js architecture you are running and the
architecture of your machine.

You will also need the boost headers on your machine.  You do not need
to build any boost libraries.

In your environment variables make sure MYSQL_DIR and BOOST_ROOT are
both set.

Now you can npm install nk-mysql and it will attempt to build.  You will
see the strange unresolved external references that are preventing me
from finishing (but I'm guessing there are some macros I need to search
for within the cmake files of the oracle managed source code for the
connector in order to resolve them, or perhaps the source files for the
build are not correct - I'm still digging through it.)

The idea is that once the connector builds properly, node-gyp will
automatically place it within the build directory of the add-on (.node
binary), and the add-on will then be built.

I've put in place holders for many of the other clauses that were not
previously implemented...they are not done and need to be finished ofc.

This provides a copy of the current state.  I will finish the build
process on windows and then update the connector properly.  Afterwards I
will move on the debian linux again.

Any help is appreciated, go right ahead and fork a copy :)

Of course, nk-mvc will provide examples for all usage of the NodaKwaeri
framework and for the NodaMySQL connector, and documentation will follow
here on GitHub

- .gyp and .gypi files added for common variables and for building of
dependency library (Yeah, learning gyp and node-gyp)

- leaving 2 copies of include due to different configurations for
different systems and architectures (basically it means there will be at
least 4 copies of the includes, will try to learn how to uncrompress
only the copy we need...saw a sneak peak of something like that in
sqlite's .gyp files)



Change-Id: I0713701b5bb22685259a1e08a0dfc17c5bf49f7f
Signed-off-by: Richard Winters <rik@mmogp.com>