Project Update
- Connector and Data Integration Tools now build on windows properly.
Please see the readme file for installation and usage instructions.
Precompiled version for windows not yet included for 32 bit systems

- Properly configured source files do not yet exist for linux, but will
be checked in within the next couple of days.

- Several clauses are not implemented correctly within the Driver and
are simply a copy of select or where clause definitions, these will be
updated shortly as well.

- The execute() method is only built to execute a specific query, it
will be updated quickly.

- Modified directory structure, added before and after build actions to
.gyp file so that proper index.js is used for module depending on
whether the module was built or not.

- Updated Readme and added package.json file.

- Setting version to 0.0.1

Change-Id: I9175fb7df5bac8ab6eced50909074bce04e47fcf
Signed-off-by: Richard Winters <rik@mmogp.com>