Project Update
- Re-added all the precompiled binaries except for windows 32-bit and
debian 64-bit, and also added the default index.js file to make use of
them.  Seems it would be necessary to force users to set the path to the
libmysql.dll file on windows, and to ensure that the respective library
is configured on linux/unix if they are not building.

- Added logic in build to copy libmysql.lib and libmysql.dll to a
directory up from node_modules - with the idea that the project making
use of them (or rather the directory which contains it) will likely be
the place where developers will invoke the application.  This is where
nk-mysql will look for libmysql.lib and libmysql.dll.  It's ok to remove
them if your system is already configured for sharing them.  They are
not generated on Linux since linux tends to configure library paths
(ldconfig on debian after mysql install - if for some reason it was not
already configured by the installer).

Change-Id: Ib7d1ab23e479113d0a5d5ec4036d3192ec8cccc7
Signed-off-by: Richard Winters <rik@mmogp.com>